+ fixed tags not saving for script-entities
